requirements.txt
blinker==1.4
click==6.7
Flask==0.12.2
Flask-Mail==0.9.1
Flask-Session==0.3.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
maxminddb==1.3.0
maxminddb-geolite2==2017.803
Werkzeug==0.12.2
